Remvove redundant code, reorg
