+++ components/SplitScreen/SplitScreen.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useMemo, useRef } from 'react'
-import styled from 'styled-components'
 import { useSpring, useTransition, animated } from 'react-spring'
 import {
   useAnimateWhenMounted,
