--- components/converter/Processing.js
@@ -1,53 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import styled from 'styled-components'
-import TransactionBadge from './TransactionBadge'
-
-import processing from './assets/loader.gif'
-
-function Processing({ isFinal, transactionHash }) {
-  return (
-    <ProcessingIn>
-      <div>
-        <img src={processing} alt="" />
-        <>
-          <p className="black">
-            Processing {isFinal ? 'last' : 'first'} transaction
-          </p>
-          <p>
-            Please wait for transaction{' '}
-            <TransactionBadge transactionHash={transactionHash} /> to be mined.
-            Do not close this window until the process is finished.
-          </p>
-        </>
-      </div>
-    </ProcessingIn>
-  )
-}
-
-Processing.propTypes = {
-  isFinal: PropTypes.bool,
-  transactionHash: PropTypes.string,
-}
-
-const ProcessingIn = styled.div`
-  margin-top: 24px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 30px;
-  text-align: center;
-  img {
-    margin-bottom: 15px;
-    height: 100px;
-  }
-  p {
-    max-width: 410px;
-    margin: 0;
-  }
-  p.black {
-    color: #000;
-  }
-`
-
-export default Processing
