+++ components/converter/TransactionBadge.js
@@ -4,23 +4,24 @@ import Anchor from '../Anchor/Anchor'
 import { shortenAddress } from 'lib/web3-utils'
 import { getEtherscanHref } from 'lib/utils'
 
-function TransactionBadge({ transactionHash }) {
+function TransactionBadge({ transactionHash, className }) {
   return (
     <div
+      className={className}
       css={`
         display: inline;
         background: #ebfafd;
         color: black;
-        border-radius: 3px;
-        padding: 4px 8px;
-        font-size: 20px;
-        line-height: 20px;
+        border-radius: 4px;
+        padding: 6px 12px;
+        font-size: 16px;
+        line-height: 1;
       `}
     >
       <Anchor
         href={getEtherscanHref(transactionHash)}
         css={`
-          color: black;
+          color: #20232c;
         `}
       >
         {shortenAddress(transactionHash)}
