+++ lib/web3-contracts.js
@@ -179,46 +179,38 @@ export function useBondingCurvePrice(amount, forwards = true) {
   return useMemo(() => ({ loading, price }), [loading, price])
 }
 
-// Convert ANT to ANJ action
-export function useOpenOrder() {
+export function useAllowance() {
   const { account } = useWeb3Connect()
   const antContract = useKnownContract('TOKEN_ANT')
-  const fundraisingContract = useKnownContract('FUNDRAISING')
-
-  const [antAddress] = getKnownContract('TOKEN_ANT')
   const [marketMakerAddress] = getKnownContract('MARKET_MAKER')
 
   return useCallback(
-    async (amount, forwards = true) => {
-      if (!antContract || !fundraisingContract) {
+    async amount => {
+      if (!antContract) {
         return false
       }
-      // Interacting with the bonding curve involves 3 transactions (or two,
-      // depending on the direction):
-      // 1. Approval (if there's not enough allowance and we're converting ANT -> ANJ)
-      // 2. Open a buy order (and wait for the transaction to be mined)
-      // 3. Claim the order (done in useClaimOrder)
 
-      // We first check for allowance if the direction is correct,
-      // and if we need more, approve
-      if (forwards) {
-        const allowance = await antContract.allowance(
-          account,
-          marketMakerAddress
-        )
-        if (allowance.lt(bigNum(amount))) {
-          try {
-            await antContract.approve(marketMakerAddress, amount)
-            // Don't wait for the approval to be mined before showing second transaction
-          } catch (err) {
-            throw new Error('User did not approve transaction')
-          }
-        }
+      const allowance = await antContract.allowance(account, marketMakerAddress)
+      if (allowance.lt(bigNum(amount))) {
+        return await antContract.approve(marketMakerAddress, amount)
+      }
+    },
+    [account, antContract, marketMakerAddress]
+  )
+}
+
+// Convert ANT to ANJ action
+export function useOpenOrder() {
+  const fundraisingContract = useKnownContract('FUNDRAISING')
+  const [antAddress] = getKnownContract('TOKEN_ANT')
+
+  return useCallback(
+    async (amount, toAnj = true) => {
+      if (!fundraisingContract) {
+        return false
       }
 
-      // Then, we open a buy or sell order. The next step will be handled after the
-      // transaction gets mined, in useClaimOrder()
-      return await (forwards
+      return await (toAnj
         ? fundraisingContract.openBuyOrder(antAddress, amount, {
             gasLimit: 650000,
             value: 0,
@@ -227,7 +219,7 @@ export function useOpenOrder() {
             gasLimit: 850000,
           }))
     },
-    [account, antAddress, antContract, fundraisingContract, marketMakerAddress]
+    [antAddress, fundraisingContract]
   )
 }
 
@@ -237,12 +229,13 @@ export function useClaimOrder() {
   const [antAddress] = getKnownContract('TOKEN_ANT')
 
   return useCallback(
-    async (openOrderTransactionHash, forwards = true) => {
+    async (openOrderTransactionHash, toAnj = true) => {
       const { blockNumber: batchId } = await ethersProvider.getTransaction(
         openOrderTransactionHash
       )
+
       // We claim the buy order, with the blockNumber of the emitted open order
-      return await (forwards
+      return await (toAnj
         ? fundraisingContract.claimBuyOrder(account, batchId, antAddress, {
             gasLimit: 500000,
           })
