+++ pages/index.js
@@ -1,14 +1,11 @@
 import React from 'react'
-import { ConverterProvider } from 'components/converter/converter-status'
-// import ConvertStepper from 'components/ConvertStepper/ConvertStepper'
+import { ConverterProvider } from 'components/Converter/converter-status'
 import ConversionForm from 'components/ConversionForm/ConversionForm'
 
 export default () => {
   return (
     <ConverterProvider>
       <ConversionForm />
-
-      {/* <ConvertStepper toAnj={true} /> */}
     </ConverterProvider>
   )
 }
